+++ src/App.js
@@ -1,20 +1,29 @@
 import React, { useState } from 'react';
 import './App.css';
 import Login from './components/login/Login';
+import Timeline from './components/timeline/Timeline';
+import Header from './components/header/Header';
 
 function App() {
-  const [logged, setLogged] = useState(false);
+    const [driver, setDriver] = useState(false);
+    const [user, setUser] = useState('');
 
-  const setLoginResponse = response => {
-    setLogged(response);
-  };
+    const setLoginResponse = response => {
+        setDriver(response.drv);
+        setUser(response.user);
+    };
+
+    const logout = () => {
+        setDriver(null);
+    }
 
 
-  return (
-    <div className="App">
-      { !logged ? <Login callback={ setLoginResponse }></Login> : null }
-    </div>
-  );
+    return (
+        <div className="App">
+            { !driver ? null : <Header user={user} callback={ logout }></Header>}
+            { !driver ? <Login callback={ setLoginResponse }></Login> : <Timeline driver={driver}></Timeline> }
+        </div>
+    );
 }
 
 export default App;